backend login desarrollo

+++ backend/src/main/java/colegio/comedor/interfaces/InterfazCriterio.java
@@ -1,12 +1,14 @@
 package colegio.comedor.interfaces;
 
+
+
+
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
 
 import colegio.comedor.modelo.Criterio;
 
 @Repository
-public interface InterfazCriterio extends  CrudRepository<Criterio,Integer>{
+public interface InterfazCriterio extends CrudRepository<Criterio,Long> {
 
 }
